# blog-it/api/swagger/swagger.yaml
openapi : 3.0.0
info:
    title: 'BlogsPage API'
    version: '1.0'
    description: 'This is the documentation for the project'
    contact: 
        name: Group 27
        email: sampthkumarreddy.m19@iiits.in
servers: 
    -   url : http://localhost:5000
        description: "main server"
    -   url : http://localhost:5000/beta
        description: "beta server"
...
paths:
...
    /api/users/{user_id}:
        get:
            summary: "Get a single user data"
            tags: [User]
            parameters:
              - in: path
                name: user_id
                required: true
                schema: 
                    type: string
            responses:
              '200' :
                description: "Single post Displayed successfully"
                content:
                  application/json:
                    schema:
                        $ref: '#/components/schemas/User'
    /api/users/sp/{user_id}:
        put:
            summary: 'Edit details of the user'
            tags: [User]
            parameters:
              - in : path
                name: user_id
                required: true
                schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                '200':
                    description: 'Success fully edited the details'
        delete:
            summary: Deletes a User 
            tags: [User]
            parameters:
              - name: user_id
                in: path
                required: true
                description: auto-generated User ID by Mongodb
                schema: 
                    type: string
            responses:
              '200' :
                description: "Deleted Successfully"
                content:
                  application/json:
                    schema:
                        $ref: '#/components/schemas/User'  
              '401':
                description: "You are not loggedin with same account"
...
components:
    schemas:
...
        User:
            type: object
            required:
                - username
                - email
            properties:
                username:
                    type: string
                email:
                    type: string
            example:
                username: sampath
                email: sampthkumarreddy.m19@iiits.in
...